# Refactor Frame reductions