… python/cudf/cudf/core/frame.py
… python/cudf/cudf/core/series.py
… python/cudf/cudf/tests/test_dataframe.py
… python/cudf/cudf/tests/test_index.py
… python/cudf/cudf/tests/test_reductions.py